requirements.txt
boto3==1.41.5
Django==5.2.4
djangorestframework==3.16.1
djlint==1.36.4
faker==38.2.0
gunicorn==23.0.0
moto[boto3]==5.1.17
psycopg2-binary==2.9.11
pydantic==2.11.7
python-dotenv==1.1.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1